fix(keycloak): stop the comment contradicting the line below it, and build on 32-bit
The doc comment said the assertions were "deliberately written in literals
rather than in terms of the constants" — and the commit before this one
replaced the literal 40s with 4*tokenRequestTimeout directly beneath it. The
comment now describes what is actually there, and distinguishes the two kinds
of bound rather than lumping them together:

The threshold stays a literal, which is what pins the value that mattered.
The window's lower bound derives from the timeout, constraining the
relationship rather than restating a value — and being the exact complement of
the compile-time guard, it cannot fire while that guard exists, so its message
will never be seen. It is a backstop for the guard being deleted: verified by
deleting the guard and setting the window to 40s, which is what makes it fire.
The upper bound is a literal and is therefore the only assertion there that can
fire while everything else is in place.

The guard converts through uint64 now. In nanoseconds its satisfied value is
about 5e9, which overflows a 32-bit uint, so it failed to build under
GOARCH=386 and GOARCH=arm — nothing targets those today, but the guard should
not be the reason a build breaks if one appears. A negative constant still
fails to convert, so it catches the same edits on both word sizes: 40s exactly
now fails with "constant -1 overflows uint64" on amd64 and on 386 alike.

+++ pkg/keycloak/client.go
@@ -109,7 +109,13 @@ const (
 // that misses its cache — changes no constant here, so the real worst case
 // would grow while this still builds. Anyone adding one has to raise the
 // multiplier by hand.
-const _ = uint(loginFailureBackoff - 4*tokenRequestTimeout - 1)
+//
+// uint64 rather than uint, because in nanoseconds the satisfied value is about
+// 5e9: that overflows a 32-bit uint, so the guard failed to build under
+// GOARCH=386 and GOARCH=arm. Nothing here targets those today, and a negative
+// constant still fails to convert, so the guard behaves identically where it
+// matters.
+const _ = uint64(loginFailureBackoff - 4*tokenRequestTimeout - 1)
 
 func (c *Client) Token() (string, error) {
 	token := c.AccessToken(context.Background())
